requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LLM Integration
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.0.26

# Environment Variables
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.7

# Additional utilities
requests==2.31.0
beautifulsoup4==4.12.2
